Add generic pointer_flow_analysis entry point
